Fix problem with date-time datatype
git-svn-id: https://factplusplus.googlecode.com/svn/trunk@1172 2c28adb8-172f-0410-82af-bdfbc96ff48c

+++ Kernel/tDataEntry.h
@@ -204,13 +204,13 @@ protected:	// methods
 		if ( typeName == "string" )
 			comp = ComparableDT(getName());
 		else if ( typeName == "number" )
-			comp = ComparableDT((long)atoi(getName()));
+			comp = ComparableDT(atol(getName()));
 		else if ( typeName == "real" )
 			comp = ComparableDT((float)atof(getName()));
 		else if ( typeName == "bool" )	// FIXME!! dirty hack
 			comp = ComparableDT(getName());
 		else if ( typeName == "time" )
-			comp = ComparableDT ( (long)atoi(getName()), 0 );
+			comp = ComparableDT ( atol(getName()), 0 );
 		else	// no more types available
 			fpp_unreachable();
 	}
